wip: rename all dbtypes refrences to point to types index, rename calculations imports
