+++ backend/src/routes/locality.ts
@@ -1,7 +1,7 @@
 import { Request, Router } from 'express'
 import { getAllLocalities, getLocalityDetails, validateEntireLocality } from '../services/locality'
 import { fixBigInt } from '../utils/common'
-import { EditDataType, EditMetaData, LocalityDetailsType } from '../../../frontend/src/shared/types/dbTypes'
+import { EditDataType, EditMetaData, LocalityDetailsType } from '../../../frontend/src/shared/types'
 import { requireOneOf } from '../middlewares/authorizer'
 import { Role } from '../../../frontend/src/shared/types'
 import { deleteLocality, writeLocality } from '../services/write/locality'
