+++ backend/src/routes/reference.ts
@@ -11,7 +11,7 @@ import {
 } from '../services/reference'
 import { requireOneOf } from '../middlewares/authorizer'
 import { Role } from '../../../frontend/src/shared/types'
-import { EditMetaData, ReferenceDetailsType, EditDataType } from '../../../frontend/src/shared/types/dbTypes'
+import { EditMetaData, ReferenceDetailsType, EditDataType } from '../../../frontend/src/shared/types'
 import { deleteReference, writeReference } from '../services/write/reference'
 
 const router = Router()
