+++ backend/src/services/crossSearch.ts
@@ -1,4 +1,4 @@
-import { User, CrossSearch } from '../../../frontend/src/shared/types/dbTypes'
+import { User, CrossSearch } from '../../../frontend/src/shared/types'
 import { Role } from '../../../frontend/src/shared/types'
 import { nowDb } from '../utils/db'
 import {
