+++ backend/src/services/localitySpeciesExport.ts
@@ -1,5 +1,5 @@
 import { PoolConnection } from 'mariadb'
-import { User } from '../../../frontend/src/shared/types/dbTypes'
+import { User } from '../../../frontend/src/shared/types'
 import { NOW_DB_NAME } from '../utils/config'
 import { Role } from '../../../frontend/src/shared/types'
 import { pool } from '../utils/db'
