+++ backend/src/services/timeBound.ts
@@ -1,6 +1,6 @@
 import { logDb, nowDb } from '../utils/db'
 import { validateTimeBound } from '../../../frontend/src/shared/validators/timeBound'
-import { TimeBoundDetailsType, EditDataType } from '../../../frontend/src/shared/types/dbTypes'
+import { TimeBoundDetailsType, EditDataType } from '../../../frontend/src/shared/types'
 import Prisma from '../../prisma/generated/now_test_client'
 import { ValidationObject } from '../../../frontend/src/shared/validators/validator'
