+++ backend/src/services/write/timeUnit.ts
@@ -4,7 +4,7 @@ import {
   TimeUnitDetailsType,
   LocalityDetailsType,
   User,
-} from '../../../../frontend/src/shared/types/dbTypes'
+} from '../../../../frontend/src/shared/types'
 import { NOW_DB_NAME } from '../../utils/config'
 import { WriteHandler } from './writeOperations/writeHandler'
 import { getFieldsOfTables } from '../../utils/db'
