+++ backend/src/utils/cascadeHandler.ts
@@ -1,4 +1,4 @@
-import { EditDataType, TimeUnitDetailsType, TimeBoundDetailsType } from '../../../frontend/src/shared/types/dbTypes'
+import { EditDataType, TimeUnitDetailsType, TimeBoundDetailsType } from '../../../frontend/src/shared/types'
 import { nowDb } from './db'
 import { calculateLocalityMaxAge, calculateLocalityMinAge } from './ageCalculator'
