+++ frontend/src/shared/types/index.ts
@@ -1,4 +1,3 @@
-
 export * from './misc';
 export * from './dbTypes';
 export * from './util';
